refac(rust): move launch arg selection logic into own file
