+++ src-tauri/src/lib.rs
@@ -44,7 +44,8 @@ mod loadouts;
 use loadouts::{
     create_loadout, delete_loadout, edit_loadout, get_all_loadout_json, get_all_loadout_names,
     get_loadout_json, get_loadout_json_as_struct, get_loadout_names, get_loadouts_path,
-    import_loadout_from_path, loadout_client_launch_args_to_vec, loadout_common_launch_args_to_vec,
+    import_loadout_from_path,
+    launch_arg_selector::{loadout_client_launch_args_to_vec, loadout_common_launch_args_to_vec},
     refresh_loadout, save_server_guid, server_key_exists, server_key_setup, start_loadout,
 };
